refactor: merge hooks for paginated batches
