+++ web-marketplace/src/pages/Dashboard/MyCreditBatches/MyCreditBatches.tsx
@@ -12,7 +12,7 @@ import { useWallet } from 'lib/wallet/wallet';
 
 import WithLoader from 'components/atoms/WithLoader';
 import { CreditBatches } from 'components/organisms';
-import { usePaginatedBatchesByIssuer } from 'hooks/batches/usePaginatedBatchesByIssuer';
+import { usePaginatedBatches } from 'hooks/batches/usePaginatedBatches';
 
 import { NO_CREDIT_BATCHES_MESSAGE } from './MyCreditBatches.constants';
 
@@ -20,7 +20,7 @@ export const MyCreditBatches = (): JSX.Element => {
   const theme = useTheme();
   const { wallet } = useWallet();
   const { batchesWithSupply, setPaginationParams, paginationParams } =
-    usePaginatedBatchesByIssuer({ address: wallet?.address });
+    usePaginatedBatches({ address: wallet?.address });
   const hasNoBatches = batchesWithSupply && batchesWithSupply?.length === 0;
 
   return (
